variable-width: Only set fixed height by default
Set player width to 100% by default, but player height to the default
proportions for a 760px width. The player height doesn't automatically
adjust to changing widths, and instead adds black lines to preserve the
video proportions. With these defaults, we now allow for browser
resizing, which can change the available width - the video will now
reduce in size, and add black lines on top/bottom to compensate for the
reduced width.

Also change variable names to video width/height to default values on
existing courses.